requirements.txt
scikit-learn==1.5.2
python-fasthtml==0.7.0
matplotlib==3.9.2
scipy==1.14.1
numpy==2.1.2
pandas==2.2.3
sqlite_minutils==4.0.3
pytest
flake8
ipython
#### Add text to install the python-package here
python-package/dist/employee_events-0.0.tar.gz
